requirements.txt
gql[all]
requests
pygnmi
pyaml